refactor: 회원/인증 페이지 구조 분리
- PageController → MemberPageController + AuthPageController로 분리
- URL 구조: /main, /my-page → /member/main, /member/my-page
- main.html → templates/member/main.html 이동
- auth 패키지에 AuthPageController 추가
- SecurityConfig: /member/** MEMBER 역할 보호, /admin/**, /master/** 페이지 라우트 추가
- 템플릿 URL 참조 업데이트 (login.html, navbar.html)
