+++ src/main/java/com/aenggukland/letspt/auth/AuthPageController.java
@@ -1,0 +1,18 @@
+package com.aenggukland.letspt.auth;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AuthPageController {
+
+    @GetMapping({"/", "/login"})
+    public String login() {
+        return "auth/login";
+    }
+
+    @GetMapping("/register")
+    public String register() {
+        return "auth/register";
+    }
+}
