+++ src/main/resources/templates/auth/login.html
@@ -214,7 +214,7 @@
                 localStorage.setItem('accessToken', data.accessToken);
                 localStorage.setItem('refreshToken', data.refreshToken);
                 showAlert('로그인 성공! 메인 페이지로 이동합니다.', true);
-                setTimeout(() => location.href = '/main', 1200);
+                setTimeout(() => location.href = '/member/main', 1200);
             } else {
                 const data = await res.json();
                 showAlert(data.message || '아이디 또는 비밀번호가 올바르지 않습니다.', false);
