+++ src/main/resources/templates/fragments/navbar.html
@@ -127,7 +127,7 @@
 <body>
 
 <header class="site-header" th:fragment="navbar">
-    <a class="logo" href="/main">LETSPT</a>
+    <a class="logo" href="/member/main">LETSPT</a>
     <nav class="nav-menu">
         <div class="nav-item">
             <a class="nav-link" href="#">트레이너 정보</a>
